Update package to v0.0.2 Add comment in all function in Localisation class

+++ lang/lang.py
@@ -1,10 +1,11 @@
 import json
 import os
+import re
 
 try:
     from local.localisation import Localisation
 except:
-    print("\n\nLang package does not work !\nPlease restart the app to finish the initialisation !\n\n")
+    print("\n\nPlease restart the app to finish the initialisation !\n\n")
 
 
 
@@ -24,35 +25,41 @@ class LangInit:
             file = open(f"./localisation/{default_app_lang}.json", "w")
             file.write("{}")
             file.close()
-        else:
-            lang_files = os.listdir("./localisation")
 
-            for l in lang_files:
-                if l.endswith(".json"):
-                    l = l.removesuffix(".json")
-                    local_json_file = open(f"./localisation/{l}.json", "rb")
-                    json_lang: dict[str, str] = json.loads(local_json_file.read())
-                    local_json_file.close()
+        if file_path_exist == False or default_lang_path_exist == False:
+            exit(0)
+        
+        lang_files = os.listdir("./localisation")
 
-                    new_local_json_file = open(f"./local/{l}.json", "w")
-                    new_local_json_file.write(json.dumps(json_lang))
-                    new_local_json_file.close()
+        for l in lang_files:
+            if l.endswith(".json"):
+                l = l.removesuffix(".json")
+                local_json_file = open(f"./localisation/{l}.json", "rb")
+                json_lang: dict[str, str] = json.loads(local_json_file.read())
+                local_json_file.close()
+
+                new_local_json_file = open(f"./local/{l}.json", "w")
+                new_local_json_file.write(json.dumps(json_lang))
+                new_local_json_file.close()
                     
-            default_local_json_file = open(f"./localisation/{default_app_lang}.json", "rb")
-            default_json_lang: dict[str, str] = json.loads(default_local_json_file.read())
-            default_local_json_file.close()
+        default_local_json_file = open(f"./localisation/{default_app_lang}.json", "rb")
+        default_json_lang: dict[str, str] = json.loads(default_local_json_file.read())
+        default_local_json_file.close()
 
-            local_py_file = open("./local/localisation.py", "w")
+        local_py_file = open("./local/localisation.py", "w")
 
-            python_lang = "import json\nclass Localisation:\n\t__lang: str\n\tdef __init__(self, lang: str) -> None:\n\t\tself.__lang = lang\n\tdef __get_local_str(self, key: str) -> str | None:\n\t\ttry:\n\t\t\tlang_file = open(f\"./local/{self.__lang}.json\", \"rb\")\n\t\t\tlang_js: dict[str, str] = json.loads(lang_file.read())\n\t\t\tlang_file.close()\n\t\t\treturn lang_js.get(key)\n\t\texcept:\n\t\t\tprint(f\"\\n\\nLocalisation {self.__lang} is not supported\\n\\n\")\n\t\t\treturn None"
+        python_lang = "import json\nclass Localisation:\n\t__lang: str\n\tdef __init__(self, lang: str) -> None:\n\t\tself.__lang = lang\n\tdef __get_local_str(self, key: str) -> str | None:\n\t\ttry:\n\t\t\tlang_file = open(f\"./local/{self.__lang}.json\", \"rb\")\n\t\t\tlang_js: dict[str, str] = json.loads(lang_file.read())\n\t\t\tlang_file.close()\n\t\t\treturn lang_js.get(key)\n\t\texcept:\n\t\t\tprint(f\"\\n\\nLocalisation {self.__lang} is not supported\\n\\n\")\n\t\t\treturn None"
 
-            if len(default_json_lang.keys()) > 0:
-                for k in default_json_lang.keys():
-                    key = k.replace(" ", "_").lower()
-                    python_lang += f"\n\tdef {key}(self):\n\t\treturn self.__get_local_str(\"{key}\")"
+        if len(default_json_lang.keys()) > 0:
+            for k in default_json_lang.keys():
+                
+                value = default_json_lang.get(k)
+                key = k.replace(" ", "_").lower()
+                
+                python_lang += f"\n\tdef {key}(self):\n\t\t\"\"\"In {default_app_lang} this message is translate to: ``{value}``\n\t\t\"\"\"\n\t\treturn self.__get_local_str(\"{key}\")"
 
-            local_py_file.write(python_lang)
-            local_py_file.close()
+        local_py_file.write(python_lang)
+        local_py_file.close()
 
     def getLocalisation(self, lang: str):
         if os.path.exists("./local/localisation.py"):
